refactored migration
