+++ pallets/messages/src/migration/mod.rs
@@ -1,0 +1,2 @@
+/// migrations to v2
+pub mod v2;
